create function to get input ft from user
