+++ src/js/views/searchView.js
@@ -1,0 +1,3 @@
+const searchInput = document.getElementById('search-input');
+
+export const getInput = () => searchInput.value;
